// Frontend/src/types/dynamic-part.ts
export type DynamicPart = {
  id: number;
  page: string;
  key: string;
  title_1?: string;
  title_2?: string;
  image_1?: string;
  image_2?: string;
  description?: string;
  link_title_1?: string;
  link_1?: string;
  link_title_2?: string;
  link_2?: string;
  created_at: Date;
  updated_at: Date;
};